refactor saving, loading, resetting

+++ src/io.c
@@ -1,5 +1,7 @@
 #include "io.h"
+#include "ui.h"
 #include "cargo.h"
+#include "render.h"
 #include "building.h"
 #include "location.h"
 #include "generate.h"
@@ -18,9 +20,27 @@ uint8_t m_worldVersions[WORLD_SAVE_SLOTS];
 
 bool m_worldExists[WORLD_SAVE_SLOTS];
 
+enum kSaveLoadRequest m_doFirst = kDoNothing, m_andThen = kDoNothing;
+
+bool m_secondStage = false;
+
+int8_t m_forceSlot = -1;
+
+uint32_t m_actionProgress = 0;
+
+SDFile* m_file;
+
 int doRead(void* _userdata, uint8_t* _buf, int _bufsize);
 
 void doWrite(void* _userdata, const char* _str, int _len);
+
+bool doSave(bool _synchronous);
+
+bool doLoad(void);
+
+bool doNewWorld(void);
+
+bool doReset(void);
 
 void decodeError(json_decoder* jd, const char* _error, int _linenum);
 
@@ -30,7 +50,60 @@ void scanDidDecode(json_decoder* jd, const char* _key, json_value _value);
 
 int scanShouldDecodeTableValueForKey(json_decoder* jd, const char* _key);
 
+json_encoder m_je;
+
+json_decoder m_jd_p = {
+  .decodeError = decodeError,
+  .didDecodeTableValue = didDecodeTableValuePlayer,
+  .didDecodeArrayValue = deserialiseArrayValuePlayer
+};
+
+json_decoder m_jd = {
+  .decodeError = decodeError,
+  .willDecodeSublist = willDecodeSublist
+};
+
+
 /// ///
+
+void doIO(enum kSaveLoadRequest _first, enum kSaveLoadRequest _andThen) {
+  m_doFirst = _first;
+  m_andThen = _andThen;
+  m_secondStage = false;
+  m_actionProgress = 0;
+}
+
+enum kSaveLoadRequest currentIOAction() {
+  return (m_secondStage ? m_andThen : m_doFirst);
+}
+
+void enactIO() {
+
+  bool finished = false;
+  switch (currentIOAction()) {
+    case kDoNothing: finished = true; break;
+    case kDoSave: finished = doSave(/*synchronous = */ false); break;
+    case kDoLoad: finished = doLoad(); break;
+    case kDoNewWorld: finished = doNewWorld(); break;
+    case kDoReset: finished = doReset(); break;
+  }
+
+  if (finished) {
+    if (!m_secondStage) {
+      m_secondStage = true;
+      m_actionProgress = 0;
+    } else {
+      // Specials...
+      m_secondStage = false;
+      m_actionProgress = 0;
+      m_doFirst = kDoNothing;
+      m_andThen = kDoNothing;
+      m_forceSlot = -1;
+      updateRenderList();
+      showTutorialMsg(getTutorialStage());
+    }
+  }
+}
 
 bool hasSaveData() { return m_foundSaveData; }
 
@@ -52,14 +125,81 @@ void setSlot(uint8_t _slot) { m_slot = _slot; }
 
 uint8_t getSlot() { return m_slot; }
 
+void setForceSlot(int8_t _slot) {
+  m_forceSlot = _slot;
+}
+
 ///
 
-void hardReset() {
-  for (uint16_t ss = 0; ss < WORLD_SAVE_SLOTS; ++ss) {
-    snprintf(m_filePath, 16, "world%u.json", (unsigned)ss);
-    pd->file->unlink(m_filePath, 0);
+bool doNewWorld() {
+  pd->system->logToConsole("NEW WORLD: Progress %i", m_actionProgress);
+  pd->sprite->addSprite(getGenSprite());
+
+  // Create and go to new world!
+  if (m_actionProgress == 0) {
+
+    reset(false); // Don't reset the player!
+    if (m_forceSlot == -1) pd->system->error("doNewWorld: must set slot override!");
+    setSlot(m_forceSlot);
+
+  } else if (m_actionProgress < 9) {
+    
+    generate(m_actionProgress);
+
+  } else if (m_actionProgress == 9) {
+
+    addObstacles();
+
+  } else if (m_actionProgress == 10) {
+
+    setChunkBackgrounds();
+    // Finished
+    // Need to save now, so issue another doIO command from here
+    doIO(kDoSave, kDoNothing);
+    return false; // Hence then have to return false otherwise we'd complete the current doIO...
   }
-  pd->file->unlink("player.json", 0);
+
+  ++m_actionProgress;
+  return false;
+}
+
+///
+
+bool doReset() {
+  pd->system->logToConsole("RESET: Progress %i", m_actionProgress);
+  pd->sprite->addSprite(getGenSprite());
+
+  if (m_actionProgress == 0) {
+
+    // Hard file system reset
+    for (uint16_t ss = 0; ss < WORLD_SAVE_SLOTS; ++ss) {
+      snprintf(m_filePath, 16, "world%u.json", (unsigned)ss);
+      pd->file->unlink(m_filePath, 0);
+    }
+    pd->file->unlink("player.json", 0);
+
+    // Internal full reset
+    reset(true);
+    setSlot(0);
+
+  } else if (m_actionProgress < 9) {
+
+    generate(m_actionProgress);
+
+  } else if (m_actionProgress == 9) {
+
+    addObstacles();
+
+  } else if (m_actionProgress == 10) {
+
+    setChunkBackgrounds();
+
+    // Finished
+    return true;
+  }
+
+  ++m_actionProgress;
+  return false;
 }
 
 ///
@@ -93,8 +233,8 @@ void scanSlots() {
     if (m_worldExists[ss]) {
       if (m_worldVersions[ss] == 0) {
         pd->system->logToConsole("Scan world: OLD WORLD DETECTED! Version 0. ACTION: Delete everything and start again");
-        hardReset();
         m_foundSaveData = false;
+        doReset();
       }
     }
   }
@@ -103,14 +243,12 @@ void scanSlots() {
     SDFile* file = pd->file->open("player.json", kFileRead|kFileReadData);
     if (!file) {
       m_foundSaveData = false;
-      pd->system->logToConsole("Scan world: No Playe data found!");
+      pd->system->logToConsole("Scan world: No Player data found!");
     }
   }
 
   pd->system->logToConsole("Scan world: overall result:%s", m_foundSaveData ? "CAN-BE-LOADED" : "CANNOT-LOAD");
 }
-
-
 
 int scanShouldDecodeTableValueForKey(json_decoder* jd, const char* _key) {
   return (strcmp(_key, "sf") == 0);
@@ -127,115 +265,145 @@ void scanDidDecode(json_decoder* jd, const char* _key, json_value _value) {
 
 ///
 
-bool save() {
+void synchronousSave() {
+  m_actionProgress = 0;
+  while (doSave(/*synchronous = */ true) == false) {
+    // noop
+  }
+}
+
+bool doSave(bool _synchronous) {
   uint8_t pretty = 0;
   #ifdef DEV
   pretty = 1;
   #endif
 
-  pd->system->logToConsole("START save player");
+  if (!_synchronous) pd->sprite->addSprite(getSaveSprite());
+  pd->system->logToConsole("SAVE: Sync:%i, Progress %i", _synchronous, m_actionProgress);
 
-  // Should get the latest export averages, guaranteed to be between 60 and 120s worth of data
-  updateExport();
+  if (m_actionProgress == 0) {
 
-  json_encoder je_p;
+    // Should get the latest export averages, guaranteed to be between 60 and 120s worth of data
+    updateExport();
 
-  snprintf(m_filePath, 32, "TMP_player.json");
-  SDFile* file_p = pd->file->open(m_filePath, kFileWrite);
+    snprintf(m_filePath, 32, "TMP_player.json");
+    m_file = pd->file->open(m_filePath, kFileWrite);
 
-  pd->json->initEncoder(&je_p, doWrite, file_p, pretty);
+    pd->json->initEncoder(&m_je, doWrite, m_file, pretty);
 
-  je_p.startTable(&je_p);
-  serialisePlayer(&je_p);
-  je_p.endTable(&je_p);
+  } else if (m_actionProgress == 1) {
 
-  int status = pd->file->close(file_p);
+    m_je.startTable(&m_je);
+    serialisePlayer(&m_je);
+    m_je.endTable(&m_je);
 
-  /// ///
+    int status = pd->file->close(m_file);
 
-  pd->system->logToConsole("START save world");
+  } else if (m_actionProgress == 2) {
 
-  json_encoder je;
+    snprintf(m_filePath, 32, "TMP_world%i.json", m_slot);
+    SDFile* file = pd->file->open(m_filePath, kFileWrite);
 
-  snprintf(m_filePath, 32, "TMP_world%i.json", m_slot);
-  SDFile* file = pd->file->open(m_filePath, kFileWrite);
+    pd->json->initEncoder(&m_je, doWrite, m_file, pretty);
 
-  pd->json->initEncoder(&je, doWrite, file, pretty);
+    m_je.startTable(&m_je);
 
-  je.startTable(&je);
+    m_je.addTableMember(&m_je, "sf", 2);
+    m_je.writeInt(&m_je, SAVE_FORMAT);
 
-  je.addTableMember(&je, "sf", 2);
-  je.writeInt(&je, SAVE_FORMAT);
+  } else if (m_actionProgress == 3) {
+    serialiseCargo(&m_je);
+  } else if (m_actionProgress == 4) {
+    serialiseBuilding(&m_je);
+  } else if (m_actionProgress == 5) {
+    serialiseLocation(&m_je);
+  } else if (m_actionProgress == 6) {
+    serialiseWorld(&m_je);
+  } else if (m_actionProgress == 7) {
+    m_je.endTable(&m_je);
 
-  serialiseCargo(&je);
-  serialiseBuilding(&je);
-  serialiseLocation(&je);
-  serialiseWorld(&je);
+    int status = pd->file->close(m_file);
+    m_file = NULL;
 
-  je.endTable(&je);
+    pd->system->logToConsole("SAVE: Saved to slot %u, save status %i", m_slot, status);
 
-  status &= pd->file->close(file);
+    // Finish by moving into location
+    char filePathFinal[32];
 
-  //float f; for (int32_t i = 0; i < 10000; ++i) for (int32_t j = 0; j < 100000; ++j) { f*=i*j; }
+    snprintf(m_filePath, 32, "TMP_player.json");
+    snprintf(filePathFinal, 32, "player.json");
+    pd->file->rename(m_filePath, filePathFinal);
 
-  pd->system->logToConsole("save to %u, status %i", m_slot, status);
+    snprintf(m_filePath, 32, "TMP_world%i.json", m_slot);
+    snprintf(filePathFinal, 32, "world%i.json", m_slot);
+    pd->file->rename(m_filePath, filePathFinal);
 
-  // Finish by moving into location
-  char filePathFinal[32];
+  } else if (m_actionProgress == 8) {
+    scanSlots();
 
-  snprintf(m_filePath, 32, "TMP_player.json");
-  snprintf(filePathFinal, 32, "player.json");
-  pd->file->rename(m_filePath, filePathFinal);
+    // Finished
+    float f; for (int32_t i = 0; i < 10000; ++i) for (int32_t j = 0; j < 100000; ++j) { f*=i*j; }
+    return true;
+  }
 
-  snprintf(m_filePath, 32, "TMP_world%i.json", m_slot);
-  snprintf(filePathFinal, 32, "world%i.json", m_slot);
-  pd->file->rename(m_filePath, filePathFinal);
-
-  return true;
+  ++m_actionProgress;
+  return false;
 }
 
 ///
 
-bool load(int32_t _forceSlot) {
+bool doLoad() {
 
-  pd->system->logToConsole("START load player");
+  pd->sprite->addSprite(getLoadSprite());
+  pd->system->logToConsole("LOAD: Progress %i", m_actionProgress);
 
-  json_decoder jd_p = {
-    .decodeError = decodeError,
-    .didDecodeTableValue = didDecodeTableValuePlayer,
-    .didDecodeArrayValue = deserialiseArrayValuePlayer
-  };
+  if (m_actionProgress == 0) {
 
-  snprintf(m_filePath, 16, "player.json");
-  SDFile* file_p = pd->file->open(m_filePath, kFileRead|kFileReadData);
+    // Clear in preparation for load (including player, if we haven't been given a slot override)
+    const bool resetThePlayerToo = (m_forceSlot == -1);
+    reset(resetThePlayerToo);
 
-  pd->json->decode(&jd_p, (json_reader){ .read = doRead, .userdata = file_p }, NULL);
+    snprintf(m_filePath, 16, "player.json");
+    m_file = pd->file->open(m_filePath, kFileRead|kFileReadData);
 
-  pd->file->close(file_p);
+    pd->json->decode(&m_jd_p, (json_reader){ .read = doRead, .userdata = m_file }, NULL);
 
-  // We have now loaded the correct slot number for this player-save.
-  // But we might want to be loading into a different world 
-  if (_forceSlot != -1) {
-    setSlot(_forceSlot);
+    pd->file->close(m_file);
+    m_file = NULL;
+
+  } else if (m_actionProgress == 1) {
+
+    // We have now loaded the correct slot number for this player-save.
+    // But we might want to be loading into a different world 
+    if (m_forceSlot != -1) {
+      pd->system->logToConsole("LOAD: Slot Override from %i to %i", m_slot, m_forceSlot);
+      setSlot(m_forceSlot);
+    }
+
+    snprintf(m_filePath, 16, "world%i.json", m_slot);
+    m_file = pd->file->open(m_filePath, kFileRead|kFileReadData);
+
+    // TODO - call this many times, each time with a different WILL DECODE
+    pd->json->decode(&m_jd, (json_reader){ .read = doRead, .userdata = m_file }, NULL);
+
+    pd->file->close(m_file);
+    m_file == NULL;
+
+  } else if (m_actionProgress == 2) {
+
+    // Things which need to run post-load
+    addObstacles();
+    doWetness();
+    setChunkBackgrounds();
+    showTutorialMsg(getTutorialStage());
+
+    // Finished
+    float f; for (int32_t i = 0; i < 10000; ++i) for (int32_t j = 0; j < 100000; ++j) { f*=i*j; }
+    return true;
   }
 
-  pd->system->logToConsole("START load from slot %i", m_slot);
-
-  json_decoder jd = {
-    .decodeError = decodeError,
-    .willDecodeSublist = willDecodeSublist
-  };
-
-  snprintf(m_filePath, 16, "world%i.json", m_slot);
-  SDFile* file = pd->file->open(m_filePath, kFileRead|kFileReadData);
-
-  pd->json->decode(&jd, (json_reader){ .read = doRead, .userdata = file }, NULL);
-
-  pd->file->close(file);
-
-  pd->system->logToConsole("STOP load from slot %i", m_slot);
-  
-  return true;
+  ++m_actionProgress;    
+  return false;
 }
 
 void willDecodeSublist(json_decoder* jd, const char* _name, json_value_type _type) {
